feat: clean & add comments to initialize.js
From looking at
https://developers.stellar.org/docs/build/guides/dapps/soroban-contract-init-template,
my guess is that people do not find the existing `initialize.js` script
easy to understand.

I've followed the lead of @CaptainPrinz here and changed the name of
`exe` to `run`, and then I added lots of comments to each function.

I also removed the `SOROBAN_NETWORK` variable, as I think it makes this
somewhat more confusing instead of more clear. This means removing the
`network.${process.env.SOROBAN_NETWORK}` shortcut, but maybe that's ok?

I also moved around things in the `.gitignore`, just so all the
specific-to-this-template stuff would be at the bottom. A new
frontend-template guide I am writing is going to mention and link to
this stuff.